libraries work but are noisier than Prototype Glove code for some reason...

+++ code/glove/test/HallEffectLibTest/lib/FingerTracking/FingerTracking.cpp
@@ -22,7 +22,7 @@ void printFingerAngles() {
 
 
 int32_t adjustMCPAbductionAngle(int32_t i){
-  int32_t angle = proto_angles[i];
+  float angle = proto_angles[i];
   float max_angle = max_angles[i];
   float min_angle = min_angles[i];
   int32_t adjusted_angle = (int32_t)((angle - min_angle)/(max_angle-min_angle) * (2*MCP_ABDUCTION_MAX));
@@ -30,7 +30,7 @@ int32_t adjustMCPAbductionAngle(int32_t i){
 }
 
 int32_t adjustMCPFlexionAngle(int32_t i){
-  int32_t angle = proto_angles[i];
+  float angle = proto_angles[i];
   float max_angle = max_angles[i];
   float min_angle = min_angles[i];
   int32_t adjusted_angle = (int32_t)((angle - min_angle)/(max_angle-min_angle) * MCP_FLEXION_MAX);
@@ -38,7 +38,7 @@ int32_t adjustMCPFlexionAngle(int32_t i){
 }
 
 int32_t adjustPIPFlexionAngle(int32_t i){
-  int32_t angle = proto_angles[i];
+  float angle = proto_angles[i];
   float max_angle = max_angles[i];
   float min_angle = min_angles[i];
   int32_t adjusted_angle = (int32_t)((angle - min_angle)/(max_angle-min_angle) * PIP_FLEXION_MAX);
@@ -46,7 +46,7 @@ int32_t adjustPIPFlexionAngle(int32_t i){
 }
 
 int32_t adjustThumbCMCAbductionAngle(int32_t i){
-  int32_t angle = proto_angles[i];
+  float angle = proto_angles[i];
   float max_angle = max_angles[i];
   float min_angle = min_angles[i];
   int32_t adjusted_angle = (int32_t)((angle - min_angle)/(max_angle-min_angle) * (2*THUMB_CMC_ABDUCTION_MAX));
@@ -54,7 +54,7 @@ int32_t adjustThumbCMCAbductionAngle(int32_t i){
 }
 
 int32_t adjustThumbCMCFlexionAngle(int32_t i){
-  int32_t angle = proto_angles[i];
+  float angle = proto_angles[i];
   float max_angle = max_angles[i];
   float min_angle = min_angles[i];
   int32_t adjusted_angle = (int32_t)((angle - min_angle)/(max_angle-min_angle) * THUMB_CMC_FLEXION_MAX);
@@ -62,7 +62,7 @@ int32_t adjustThumbCMCFlexionAngle(int32_t i){
 }
 
 int32_t adjustThumbPIPFlexionAngle(int32_t i){
-  int32_t angle = proto_angles[i];
+  float angle = proto_angles[i];
   float max_angle = max_angles[i];
   float min_angle = min_angles[i];
   int32_t adjusted_angle = (int32_t)((angle - min_angle)/(max_angle-min_angle) * THUMB_PIP_FLEXION_MAX);
